# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: E5-ATLASIS API
  version: 0.1.0
  description: Information Security Asset & Infrastructure System
servers:
  - url: /api/v1
paths:
...
  /assets:
    post:
      operationId: createInformationAsset
      summary: Create an Information Asset
      parameters:
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInformationAssetRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: Information Asset created
          headers:
            Location:
              description: URL of the created Information Asset
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationAsset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /assets/{assetId}:
    get:
      operationId: getInformationAsset
      summary: Get an Information Asset
      parameters:
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/AssetId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Information Asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationAsset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests:
    post:
      operationId: createChangeRequest
      summary: Create a Change Request
      parameters:
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChangeRequestRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: Change Request created
          headers:
            Location:
              description: URL of the created Change Request
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}:
    get:
      operationId: getChangeRequest
      summary: Get a Change Request
      parameters:
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/ChangeRequestId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Change Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/submit:
    post:
      operationId: submitChangeRequest
      summary: Submit a Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Change Request submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/review:
    post:
      operationId: startChangeRequestReview
      summary: Start Change Request review
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Change Request is under review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/request-changes:
    post:
      operationId: requestChangeRequestChanges
      summary: Request changes to a Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Changes requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/reject:
    post:
      operationId: rejectChangeRequest
      summary: Reject a Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Change Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/approve:
    post:
      operationId: approveChangeRequest
      summary: Approve a Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Change Request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/apply:
    post:
      operationId: applyChangeRequest
      summary: Apply an approved Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Updated Information Asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationAsset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /change-requests/{changeRequestId}/approvals:
    get:
      operationId: listChangeRequestApprovals
      summary: List Change Request approvals
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createChangeRequestApproval
      summary: Create an approval for a Change Request
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: Approval created
          headers:
            Location:
              description: URL of the created approval
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /approvals/{approvalId}/approve:
    post:
      operationId: approveApproval
      summary: Approve an approval
      parameters:
        - $ref: '#/components/parameters/ApprovalId'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Approval approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /approvals/{approvalId}/reject:
    post:
      operationId: rejectApproval
      summary: Reject an approval
      parameters:
        - $ref: '#/components/parameters/ApprovalId'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Approval rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    ActorId:
      name: X-Actor-ID
      in: header
      required: true
      description: Development-only actor placeholder accepted only when ATLASIS_AUTH_MODE=development-header and HTTP_ADDR binds to a loopback IP; this header is not authentication.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    Unauthorized:
      description: Authentication is required or the principal is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    CreateInformationAssetRequest:
      type: object
      additionalProperties: false
      required:
        - type
        - name
        - organizationId
        - ownerId
        - criticality
        - riskLevel
        - security
      properties:
        type:
          type: string
          enum:
            - information_system
            - information_infrastructure_object
        name:
          type: string
          minLength: 1
          maxLength: 500
        shortName:
          type: string
          maxLength: 255
        organizationId:
          type: string
          minLength: 1
          maxLength: 128
        ownerId:
          type: string
          minLength: 1
          maxLength: 128
        purpose:
          type: string
          maxLength: 4000
        criticality:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        riskLevel:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        security:
          $ref: '#/components/schemas/SecurityProfile'
...